+++ src/consts.js
@@ -1,36 +1,3 @@
-import { getRandomIntInclusive } from './util.js';
-
-const getUserLevel = function () {
-  const randomLevel = getRandomIntInclusive(0, 100);
-  switch (Boolean(randomLevel)) {
-    case randomLevel >= 1 && randomLevel <= 10:
-      return `novice`;
-    case randomLevel >= 11 && randomLevel <= 20:
-      return `fan`;
-    case randomLevel >= 21:
-      return `movie buff`;
-    case randomLevel === 0:
-      return ``;
-    default:
-      throw new Error(`Неизвестный уровень: «` + randomLevel + `»`);
-  }
-};
-
-const monthNames = [
-  `January`,
-  `February`,
-  `March`,
-  `April`,
-  `May`,
-  `June`,
-  `July`,
-  `August`,
-  `September`,
-  `October`,
-  `November`,
-  `December`,
-];
-
 
 const commentImgNames = [
   `./images/emoji/angry.png`,
@@ -64,8 +31,6 @@ const commentDates = [
 ];
 
 export {
-  getUserLevel,
-  monthNames,
   commentImgNames,
   commentText,
   commentAuthors,
